Tidy up StringPiece and make it use size_t.
Reviewed-on: https://code-review.googlesource.com/6972

+++ re2/bitstate.cc
@@ -289,7 +289,7 @@ bool BitState::TrySearch(int id0, const char* p0) {
             (longest_ && p > submatch_[0].end())) {
           for (int i = 0; i < nsubmatch_; i++)
             submatch_[i].set(cap_[2*i],
-                             static_cast<int>(cap_[2*i+1] - cap_[2*i]));
+                             static_cast<size_t>(cap_[2*i+1] - cap_[2*i]));
         }
 
         // If going for first match, we're done.
